feat: its working
okay so we have a lot of stuff working. `just dev` deploys the whole
thing to my localhost. a cloudflare tunnel is created. docker swarm
stacks are deployed, and a service polls for service updates to services
in the swarm to update the cloudflare tunnel services. still need to
check to make sure it actually works (dealing with universal certificate
issue stuff rn). then we gotta lock some stuff behind access then WE ARE
GOLDEN.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>