feat: ermmm okay so its working?!?!?
welp... its working. and thats super cool. we got all the whole service
discovery and everything and it syncs it with the cloudflare tunnel and
with cloudflare access. really cool stuff. next up is getting authelia
up and running and working as the idp for cloudflare access. once that's
done, we can look to adding jellyfin and the starr stuff.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>