feat: create authelia idp in cloudflare
okay so holy shit this was annoying. kept running into junk with the
openssl key because cryptography kept breaking because of the stupid dev
env. dont get me wrong, i love that stuff but sometimes its just broken
or tools dont support it well or whatever. whatever. IT WORKS NOW
THOUGH!

so next up is of course making the authelia idp the default for apps
protected by cloudflare access. the only one that kinda shouldnt be
authelia is the ldap user manager, so on initial setup i can access it,
since authelia obviously wont be configured yet. apart from that, we are
basically scott free or whatever the phrase is. everything is setup that
i need so far, like application auth and whatnot. i could setup jellyfin
RIGHT NOW and it would like work. we'll get there though. need the
infrastructure there before i want to start building out the other junk.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>