feat: use authelia by default
so this was a lot easier that i though but i ran into a bunch of
consistency issues. it works but i have a bunch of state that is in
various states of working. next up is refactoring cfzt-sync to do better
consistency. probably renaming it first then setting it up to ensure
that all the apps and things have the proper configs.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>