feat: add sonarr, prowlarr, and transmission
so lots of work in this commit. but it was needed cuz this was actually
a lot of work :D basically, we got sonarr and prowlarr running, getting
automatically connected up with the manager, and transmission also
getting connected up and running through a vpn. not much is configurable
yet but im just trying to get things running.

so things get auto configured but not all the way. there are still some
things i need to configure manually which need to be added to the
manager, but we'll get there.

also still need to set up radarr, readarr, and lidarr. but now that the
manager stuff is mostly setup so it should be pretty easy. also
jellyseer or whatever its called. probably gonna do that next.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>