feat: authelia config to docker config
so, gonna start moving a bunch of configs to docker configs. big thing
here is if i dont, services that use volumes to mount config will be
tied to the machine they're running on, meaning no horizontal scaling.
alright so this allows scaling across multiple machines and better
support for all that.

gonna do this for every service, but ultimately probably gonna manually
deploy the network to the server manually and gonna set up configurable
domains in the configurer. so dev will be hosted on a different
subdomain than the production system (probably the right way of doing
things). first configs tho

Signed-off-by: Matt Provost <mattprovost6@gmail.com>