feat: add qol to cloudflared
first in a series of commits to add all these things to the services.
healthchecks and all that. spent some time working on zig but having
some really annoying errors with building it locally. gonna have to do
all that in a container or vm. that'll be a fun project.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>