feat: add prometheus
yipee we're here. get to start working with a bunch of the fun stuff.
gonna add all the metrics servers we have then get to adding grafana.
then we get to work with all that graph junk. very fun stuff. after
grafana, some sort of alerting system and gotify. gotta also set
jellyseerr to notify to gotify. lots of work ahead of us.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>